Allow emulated startup of launcher and installer
Restructured some bits to make it easier to fill in the blanks. With these
changes the emulator can successfully start the launcher and the installer,
though actually interacting with them is not yet supported.

Additional benefit is that this allows moving some of the emulation code out
of the micropython repo, keeping that repo more in sync with upstream.

Merge https://github.com/HackerHotel/micropython-esp32/pull/1 and update the
submodule in this PR before merging this PR.